Extract `traverseDoc`
