+++ src/document/utils.js
@@ -18,6 +18,7 @@ import {
 } from "./constants.js";
 import { literalline, join } from "./builders.js";
 import getDocType from "./utils/get-doc-type.js";
+import traverseDoc from "./utils/traverse-doc.js"
 import InvalidDocError from "./invalid-doc-error.js";
 
 const getDocParts = (doc) => {
@@ -32,60 +33,6 @@ const getDocParts = (doc) => {
 
   return doc.parts;
 };
-
-// Using a unique object to compare by reference.
-const traverseDocOnExitStackMarker = {};
-
-function traverseDoc(doc, onEnter, onExit, shouldTraverseConditionalGroups) {
-  const docsStack = [doc];
-
-  while (docsStack.length > 0) {
-    const doc = docsStack.pop();
-
-    if (doc === traverseDocOnExitStackMarker) {
-      onExit(docsStack.pop());
-      continue;
-    }
-
-    if (onExit) {
-      docsStack.push(doc, traverseDocOnExitStackMarker);
-    }
-
-    if (
-      // Should Recurse
-      !onEnter ||
-      onEnter(doc) !== false
-    ) {
-      // When there are multiple parts to process,
-      // the parts need to be pushed onto the stack in reverse order,
-      // so that they are processed in the original order
-      // when the stack is popped.
-      if (Array.isArray(doc) || doc.type === DOC_TYPE_FILL) {
-        const parts = getDocParts(doc);
-        for (let ic = parts.length, i = ic - 1; i >= 0; --i) {
-          docsStack.push(parts[i]);
-        }
-      } else if (doc.type === DOC_TYPE_IF_BREAK) {
-        if (doc.flatContents) {
-          docsStack.push(doc.flatContents);
-        }
-        if (doc.breakContents) {
-          docsStack.push(doc.breakContents);
-        }
-      } else if (doc.type === DOC_TYPE_GROUP && doc.expandedStates) {
-        if (shouldTraverseConditionalGroups) {
-          for (let ic = doc.expandedStates.length, i = ic - 1; i >= 0; --i) {
-            docsStack.push(doc.expandedStates[i]);
-          }
-        } else {
-          docsStack.push(doc.contents);
-        }
-      } else if (doc.contents) {
-        docsStack.push(doc.contents);
-      }
-    }
-  }
-}
 
 function mapDoc(doc, cb) {
   // Avoid creating `Map`
